PowerBI_DonutChart_Anlalysis
PowerBI analysis has been done in this assignments with the help of Donut chart, Slicers, Filters and Map visualization. In this Assignment, we have analyzed the number of complaints Region-wise and State-wise and we have also categorized the complains with the help of slicer so that we can see number of complaints region-wise. We can also see complaints and its count with the names of our Customers so that we can find out that which customer is facing what type of issue.